+++ src/mako/http/response/senders/File.php
@@ -273,9 +273,9 @@ class File implements ResponseSenderInterface
 
 		$response->type($this->getContenType());
 
-		$response->header('Accept-Ranges', $request->isSafe() ? 'bytes' : 'none');
+		$response->getHeaders()->add('Accept-Ranges', $request->isSafe() ? 'bytes' : 'none');
 
-		$response->header('Content-Disposition', $this->getDisposition() . '; filename="' . $this->getName() . '"');
+		$response->getHeaders()->add('Content-Disposition', $this->getDisposition() . '; filename="' . $this->getName() . '"');
 
 		// Get the requested byte range
 
@@ -304,7 +304,7 @@ class File implements ResponseSenderInterface
 
 				$range = ['start' => 0, 'end' => $this->fileSize - 1];
 
-				$response->header('Content-Length', $this->fileSize);
+				$response->getHeaders()->add('Content-Length', $this->fileSize);
 			}
 			else
 			{
@@ -313,9 +313,9 @@ class File implements ResponseSenderInterface
 
 				$response->status(206);
 
-				$response->header('Content-Range', sprintf('bytes %s-%s/%s', $range['start'], $range['end'], $this->fileSize));
+				$response->getHeaders()->add('Content-Range', sprintf('bytes %s-%s/%s', $range['start'], $range['end'], $this->fileSize));
 
-				$response->header('Content-Length', $range['end'] - $range['start'] + 1);
+				$response->getHeaders()->add('Content-Length', $range['end'] - $range['start'] + 1);
 			}
 
 			// Send headers and the requested byte range